Rename to `TickerVectorQueryTable`
