+++ src/components/TickerVectorQueryTable/index.tsx
@@ -1,0 +1,6 @@
+import TickerVectorQueryTable, {
+  VectorSimiliartyTableProps,
+} from "./TickerVectorQueryTable";
+
+export default TickerVectorQueryTable;
+export type { VectorSimiliartyTableProps };
