--- src/components/VectorQueryTable/index.tsx
@@ -1,6 +1,0 @@
-import VectorQueryTable, {
-  VectorSimiliartyTableProps,
-} from "./VectorQueryTable";
-
-export default VectorQueryTable;
-export type { VectorSimiliartyTableProps };
